pyproject.toml
[project]
name = "erp-fenix-soft"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.3",
    "django-allauth>=65.15.0",
    "django-environ>=0.13.0",
    "django-htmx>=1.27.0",
    "django-simple-history>=3.11.0",
    "django-tailwind>=4.4.2",
    "django-unfold>=0.86.1",
    "djangorestframework>=3.17.0",
    "psycopg2-binary>=2.9.11",
]
